+++ app/models/dockets/legacy_docket.rb
@@ -25,7 +25,7 @@ class LegacyDocket < Docket
   end
 
   def weight
-    count(priority: false) + nod_count * CaseDistributionLever.find_by_item('nod_adjustment').try(:value).to_i
+    count(priority: false) + nod_count * CaseDistributionLever.find_by_item('nod_adjustment').try(:value).to_f
   end
 
   def ready_priority_appeal_ids
